Middle of STRIPE implementation
The form has all the fields and stripe is correctly implemented to
perform credit card formatting validation. However, there are two
issues: the form does not appear to execute form_action.php when
everything is formatted correctly and the backend stripe stuff is not
working because the stripe_php library is not in the correct location (I
followed the instructions on the stripe website  for how to do this, but
they didn't work with this).